pyproject.toml
[project]
name = "monkeyking"
version = "0.1.0"
description = "个人生活助理 Agent"
readme = "README.md"
authors = [
    { name = "bytedance", email = "bytedance@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai",
    "python-dotenv",
    "langchain",
    "pydantic",
    "requests",
    "loguru",
    "typer",
    "rich>=14.3.3",
    "langchain-openai>=1.1.10",
    "prompt-toolkit>=3.0.52",
]

[project.scripts]
monkeyking = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
